+++ packages/gamut-icons/CHANGELOG.md
@@ -3,6 +3,24 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [9.28.0](https://github.com/Codecademy/gamut/compare/@codecademy/gamut-icons@9.27.1...@codecademy/gamut-icons@9.28.0) (2024-06-12)
+
+### Features
+
+- **Icons:** add ProfessionalCertificateIcon ([b58018f](https://github.com/Codecademy/gamut/commit/b58018f85ecc8dd768d8a3cc711348c53543377d))
+
+### [9.27.1](https://github.com/Codecademy/gamut/compare/@codecademy/gamut-icons@9.27.0...@codecademy/gamut-icons@9.27.1) (2024-06-12)
+
+### Bug Fixes
+
+- update Java, JS, and Python icons to latest streamline versions ([c12a25e](https://github.com/Codecademy/gamut/commit/c12a25e68ab418a50c2004df947f49cd2119c891))
+
+## [9.27.0](https://github.com/Codecademy/gamut/compare/@codecademy/gamut-icons@9.26.0...@codecademy/gamut-icons@9.27.0) (2024-06-10)
+
+### Features
+
+- add new Skills icons ([#2884](https://github.com/Codecademy/gamut/issues/2884)) ([afe8fe8](https://github.com/Codecademy/gamut/commit/afe8fe83ef084768b6f35654e2ced9db8247465c))
+
 ## [9.26.0](https://github.com/Codecademy/gamut/compare/@codecademy/gamut-icons@9.25.0...@codecademy/gamut-icons@9.26.0) (2024-06-05)
 
 ### Features